#35 updating the test applications to use different package names to enable deployment of all test apps to a given device
