+++ RendererTester/pom.xml
@@ -33,10 +33,19 @@
     
 
 	  <dependency>
-	       <groupId>${project.parent.groupId}</groupId>
+	       <groupId>sec.web.renderer</groupId>
 	       <artifactId>mil-sym-android-renderer</artifactId>
-	       <version>${project.parent.version}</version>
+	       <version>0.0.65</version>
 	       <type>aar</type>
+	  </dependency>
+	  
+	  
+	    <dependency>
+	       <groupId>sec.web.renderer</groupId>
+	       <artifactId>mil-sym-android-renderer</artifactId>
+	       <version>0.0.65</version>
+	       <type>jar</type>
+	       <scope>provided</scope>
 	  </dependency>
      </dependencies>
