+++ mapfragment-demo/src/main/java/armyc2/c2sd/renderer/test2/MyView.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.example;
+package armyc2.c2sd.renderer.test2;
 
 import android.content.Context;
 import android.graphics.Canvas;
