+++ pom.xml
@@ -15,6 +15,7 @@
   
 
      <properties>
+         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	  <it-plugin.version>4.3.0</it-plugin.version>
 	  <android.lib>4.4.2_r4</android.lib>
      </properties>
